+++ SynctreeRFC/src/main/java/com/synctree/dto/InsureDTO.java
@@ -2,7 +2,6 @@ package com.synctree.dto;
 
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.Map;
 
 public class InsureDTO {
