+++ SynctreeRFC/src/main/java/com/synctree/framework/rfc/RfcDTO.java
@@ -1,0 +1,15 @@
+package com.synctree.framework.rfc;
+
+import java.util.HashMap;
+
+public class RfcDTO {
+
+		//AES256 암복호화
+		public String str1;
+		public String str2;
+		
+		//보험료 조회
+		public HashMap<String, Object> req;
+
+		
+}
