--- SynctreeRFC/src/main/java/com/synctree/rfc/RemoteFunction.java
@@ -1,9 +1,0 @@
-package com.synctree.rfc;
-
-import org.json.simple.JSONObject;
-
-public interface RemoteFunction {
-	
-	JSONObject remoteFunction(JSONObject obj);
-	
-}
