+++ SynctreeRFC/src/main/resources/application.properties
@@ -23,3 +23,5 @@ mybatis.type-aliases-package=classpath:com.synctree.dto
 
 #log level setting
 logging.level.root=INFO
+
+server.port=8070
